requirements.txt
Django==4.2.3
python-dotenv~=1.0.0
djangorestframework~=3.14.0
django-cors-headers~=4.2.0
djangorestframework-simplejwt~=5.2.2
django-extensions~=3.2.1
psycopg2~=2.9.5